Migrate layout to server
