+++ next/app/layout.jsx
@@ -1,16 +1,6 @@
-
-'use client'
-
 import React from "react";
 
-import * as Sentry from '@sentry/nextjs';
-import { createStore, applyMiddleware, compose } from 'redux';
-import logger from 'redux-logger';
-import rootReducer from '/src/reducers';
-import { Provider } from 'react-redux';
-import ScrollToTop from '/src/components/ScrollToTop';
-import Nav from '/src/components/Nav';
-
+import { Suspense } from 'react';
 
 import '/src/styles/index.css';
 import '/src/styles/footer.css';
@@ -21,16 +11,9 @@ import '/src/styles/cart.css';
 import '/src/styles/checkout.css';
 import '/src/styles/complete.css';
 import '/src/styles/product.css';
-import Footer from "/src/components/Footer";
+
 import SentryQueryInitializer from "../ui/sentry-query-initializer";
-
-
-const sentryReduxEnhancer = Sentry.createReduxEnhancer({});
-
-const store = createStore(
-  rootReducer,
-  compose(applyMiddleware(logger), sentryReduxEnhancer)
-);
+import HomeContent from "../ui/home-content";
 
 
 export default function RootLayout({
@@ -39,16 +22,18 @@ export default function RootLayout({
   children,
 }) {
   return (
-      <html lang="en">
-        <body id="body-container">
-        <SentryQueryInitializer />
-          <Provider store={store}>
-            <ScrollToTop />
-            <Nav />
+    <html lang="en">
+      <body id="body-container">
+        <Suspense>
+          <SentryQueryInitializer />
+        </Suspense>
+        <Suspense>
+          <HomeContent>
             {children}
-          </Provider>
-          <Footer />
-        </body>
-      </html>
+          </HomeContent>
+        </Suspense>
+
+      </body>
+    </html>
   )
 }
