+++ next/sentry.client.config.js
@@ -19,7 +19,7 @@ Sentry.init({
   // Add optional integrations for additional features
   integrations: [
     Sentry.replayIntegration(),
-    Sentry.browserTracingIntegration(),
+    Sentry.browserTracingIntegration()
   ],
 
   // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
