+++ next/sentry.server.config.js
@@ -4,11 +4,14 @@
 
 import * as Sentry from '@sentry/nextjs';
 
+
 Sentry.init({
   dsn: process.env.NEXT_PUBLIC_DSN,
 
   // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
   tracesSampleRate: 1,
+
+  integrations: [Sentry.prismaIntegration()],
 
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,
