+++ next/src/components/ScrollToTop.jsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
